// sourcecode/BTrak.Packages/snova-trading-module/trading-module-components/trading-components/src/lib/billing/models/trade-template-model.ts
export class TradeTemplateModel {
    tradeTemplateName: string;
    formBgColor: string;
    tradeTemplateFormJson: string;
    tradeTemplateId: string;
    termsAndConditionId: string;
    templateTypeId: string;
    templateTypeName: string;
    isArchived: boolean;
    timeStamp: any;
    fields: any;
    formKeys : any;
}